Removing randomKey from CryptoUtil
The method randomKey used SecureRandom to generate a generate a random
seed. This call uses depends on the operating system for randomness and
was causing long stalls on linux. The method itself was only used in one
place and by unit tests. The tests can all use a hard coded tokenKey as
for the tests this does not matter.