Fixing SockJS gretting message bug.
The first access to a SockJS server, for example
http://localhost:7777/simplepush would correctly return the SockJS
welcome message. But successive calls would not return any message. This
was because the Greeting class is using an unreleasable buffer but not
duplicating it (reseting its readerIndex). When the buffer was passed
through the HttpResponseEncoder the readerIndex would be incremented and
the next time though equal to the writerIndex and an empty string would
be returned.

Updating Greetings unreleasable buffer to be duplicated upon called
fixes this.

[https://issues.jboss.org/browse/AGSMPLPUSH-49]