Improving logging for server.
Motivation:
Currently, even when debug log level is set there are now logs showing
notifications being sent to connected clients. This would improve
testing to see that messages are reaching the server.

Modifications:
Logging has been added to the NotifiacationHandler and will log the
UserAgent ID as well as the updates for the notification.

By default logging is turned on for when running the server using but
can be turned off by updating
src/main/resources/simplelogger.properties.

Logging for tests are turned off by default to avoid extra noice but can
be turned on by updating:
src/test/resources/simplelogger.properties

[https://issues.jboss.org/browse/AGSMPLPUSH-65]