Making CryptoUtil Base64 encoding url safe.
Motivation:
CryptoUtil currently version delegates its Base64 encoding to bounty
castle which does not do a url safe encoding. Url safe in this case means
that it does not emit '+' and '/'. In our case, the the encoded output is
used in an url which causes issues with some web servers that disallow
or can misinterprete them.

For example if a '/' is url encoded into '%2F' this request can be
rejected by default on some servers (like Apache httpd).

Modifications:
Changed CryptoUtil to use commons.codec's Base64 which provides a url
safe method.

Result:
The endpoint url returned when registering a channel will no longer
contain any '$2F' characters.

[https://issues.jboss.org/browse/AGSMPLPUSH-64]